% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_risque.R
\name{ajouter_zeros_tableau}
\alias{ajouter_zeros_tableau}
\title{Ajoute les croisements nuls à un tableau de données}
\usage{
ajouter_zeros_tableau(tableau, cat_vars = NULL, cnt_var = "nb_obs")
}
\arguments{
\item{tableau}{data.frame}

\item{cat_vars}{vecteur de chaines de caractères: variables de `tableau` qui sont croisées}

\item{cnt_var}{chaine de caractères: nom de la variable de comptage}
}
\value{
data.frame avec le croisement de toutes les modalités (y compris nuls)
}
\description{
Ajoute les croisements nuls à un tableau de données
}
\examples{
library(dplyr)

data("dtest")
cat_vars <- c("DEP", "DIPLOME", "SEXE", "AGE")
tab_comptage <- tabulate_cnt_micro_data(
  df = dtest, rk = NULL,
  cat_vars = cat_vars,
  marge_label = "Total"
)

avec_zeros <- ajouter_zeros_tableau(tab_comptage, cat_vars)
}
\keyword{internal}
