% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construire_table_parametres.R
\name{build_parameters_table}
\alias{build_parameters_table}
\title{Generate parameter combinations table}
\usage{
build_parameters_table(Ds, Vs, jss = 0)
}
\arguments{
\item{Ds}{Integer vector. Deviation values}

\item{Vs}{Numeric vector. Variance values}

\item{jss}{Integer vector. Sensitivity thresholds}
}
\value{
data.frame of parameter combinations
}
\description{
Creates a data.frame of all possible parameter combinations for CKM analysis.
}
\examples{
build_parameters_table(c(10,15), c(10,20), js = 5)
}
