% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrice_transition.R
\name{build_stacked_matrix}
\alias{build_stacked_matrix}
\title{Build a matrix as described in the following paper [10.13140/RG.2.2.19275.55840](10.13140/RG.2.2.19275.55840)}
\usage{
build_stacked_matrix(D, V, js, D1, V1, js1 = 0)
}
\arguments{
\item{D1}{Deviation for small counts}

\item{V1}{Variance for small counts}

\item{js1}{js for small counts}

\item{D2}{Deviation for last count}

\item{V2}{Variance for last count}

\item{js2}{js for last count}
}
\value{
ptable object
}
\description{
Build a matrix as described in the following paper [10.13140/RG.2.2.19275.55840](10.13140/RG.2.2.19275.55840)
}
\examples{
build_stacked_matrix(15, 20, 5, 15, 5, 0)
}
\references{
\insertRef{Jamme_2026}{ckm}
}
