% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimer_intervalle_confiance_ratio.R
\name{calculer_beta_ratio}
\alias{calculer_beta_ratio}
\title{Estime le beta minimal tel que P(|R-R'|>beta) < alpha, pour alpha fixé et
pour un ratio R=A/B fourni en entrée.}
\usage{
calculer_beta_ratio(
  A,
  B,
  fun = function(a, b) {
     a/b * 100
 },
  D,
  V,
  js = 0,
  ptab,
  beta_min = 0,
  beta_max = 10,
  precision = 0.1,
  alpha = 0.05,
  posterior = FALSE
)
}
\arguments{
\item{A}{numérateur}

\item{B}{dénominateur}

\item{fun}{function, sert à calculer la statistique souhaitée
à partir de A et B}

\item{D}{Déviation max de la CKM}

\item{V}{Variance de la CKM}

\item{js}{Seuil interdiction de la CKM}

\item{ptab}{data.frame issu de ptable}

\item{beta_min}{integer, seuil de précision minimal}

\item{beta_max}{integer, seuil de précision maximal}

\item{precision}{double, niveau de précision de l'estimation de beta}

\item{alpha}{double, niveau d'erreur fixé}

\item{posterior}{Booléen: Si \code{TRUE}, c'est beta' qui est calculé.}
}
\value{
numeric value
}
\description{
Estime le beta minimal tel que P(|R-R'|>beta) < alpha, pour alpha fixé et
pour un ratio R=A/B fourni en entrée.
}
\details{
Si \code{posterior=FALSE}, le calcul repose sur l'approche a priori,
c'est-à-dire que le ratio fourni (A/B) est le ratio réel/original.
Sinon, le calcul repose sur l'approche a posteriori,
c'est-à-dire que le ratio fourni (A/B) est le ratio issu de la perturbation
par CKM.

Le meilleur beta est recherché dans l'intervalle [beta_min; beta_max].
}
\examples{
library(dplyr)
ptab <- ptable::create_cnt_ptable(D = 15, V = 30.1, js = 0)@pTable |>
  as.data.frame()
r1 <- calculer_proba_ecart_ratio(
  A=100,B=1500,D = 15, V = 30.1, js = 0, ptab = ptab, betas = seq(0,10,0.1)
)
r1 |> filter(proba <= 0.05) |> head(1) |> pull(beta)
calculer_beta_ratio(A=100,B=1500,D = 15, V = 30.1, js = 0, ptab = ptab)
r2 <- calculer_proba_ecart_ratio(
  A=100,B=1500,
  D = 15, V = 30.1, js = 0,
  ptab = ptab,
  betas = seq(0,10,0.1),
  posterior = TRUE
)
r2 |> filter(proba <= 0.05) |> head(1) |> pull(beta)
calculer_beta_ratio(
  A=100,B=1500,
  D = 15, V = 30.1, js = 0,
  ptab = ptab,
  posterior = TRUE
 )
}
