% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_risque.R
\name{calculer_ensemble_deviation}
\alias{calculer_ensemble_deviation}
\title{Calcule l'ensemble de déviation d'une valeur originale i étant donnés les paramètres D et js
de la matrice de transition}
\usage{
calculer_ensemble_deviation(i, D, js = 0)
}
\arguments{
\item{i}{valeur originale dont on souhaite calculer l'ensemble de déviation}

\item{D}{la déviation de la CKM}

\item{js}{Maximum des valeurs interdites après perturbation}
}
\value{
vecteur des valeurs perturbées possibles si la valeur originale est i,
calculé en fonction de D et js.
}
\description{
Calcule l'ensemble de déviation d'une valeur originale i étant donnés les paramètres D et js
de la matrice de transition
}
\examples{

calculer_ensemble_deviation(1, 5) #attendu 0:6
calculer_ensemble_deviation(1, 5, 2) #attendu c(0,3:6)
calculer_ensemble_deviation(0, 5, 2) #attendu 0
calculer_ensemble_deviation(5, 5, 2) #attendu c(0,3:10)
}
