% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimer_intervalle_confiance_ratio.R
\name{calculer_proba_ecart_ratio}
\alias{calculer_proba_ecart_ratio}
\title{Calcule P(|R-R'|>beta) pour R= A/B et beta fixés.}
\usage{
calculer_proba_ecart_ratio(
  A,
  B,
  D,
  V,
  js = 0,
  ptab,
  betas = c(0, 1, 5, 10, 20),
  posterior = FALSE
)
}
\arguments{
\item{A}{numérateur}

\item{B}{dénominateur}

\item{D}{Déviation max de la CKM}

\item{V}{Variance de la CKM}

\item{js}{Seuil interdiction de la CKM}

\item{ptab}{data.frame issu de ptable}

\item{betas}{Seuil des écarts entre R et R'}

\item{posterior}{Booléen: Si \code{TRUE}, c'est beta' qui est calculé.}
}
\value{
dataframe de deux colonnes (beta, proba)
}
\description{
Calcule P(|R-R'|>beta) pour R= A/B et beta fixés.
}
\details{
La fonction prend un seul ratio en entrée.
Si \code{posterior=FALSE}, le calcul repose sur l'approche a priori,
c'est-à-dire que le ratio fourni (A/B) est le ratio réel/original.
Sinon, le calcul repose sur l'approche a posteriori,
c'est-à-dire que le ratio fourni (A/B) est le ratio issu de la perturbation
par CKM.
}
\examples{
ptab <- ptable::create_cnt_ptable(D = 15, V = 30.1, js = 0)@pTable |> as.data.frame()
# Calcul de alpha pour la stat originale = 10/15 pour différents beta
calculer_proba_ecart_ratio(A=10,B=15,D = 15, V = 30.1, js = 0, ptab = ptab)
# Calcul de alpha pour la stat perturbée = 10/15 pour différentes valeurs de beta'
calculer_proba_ecart_ratio(A=10,B=15,D = 15, V = 30.1, js = 0, ptab = ptab, posterior = TRUE)
}
