% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimer_intervalle_confiance_ratio.R
\name{calculer_proba_ecart_ratio_sur_df}
\alias{calculer_proba_ecart_ratio_sur_df}
\title{Calcule P(|R-R'|>beta) pour chaque R=A/B d'un dataframe donné et pour
chaque beta fixé.}
\usage{
calculer_proba_ecart_ratio_sur_df(
  data,
  fun = function(a, b) {
     a/b * 100
 },
  D,
  V,
  js = 0,
  betas = c(0, 1, 2, 5, 10, 20),
  posterior = FALSE,
  parallel = FALSE,
  max_cores = NULL
)
}
\arguments{
\item{data}{data.frame à 2 colonnes, le numérateur et le dénominateur
de chaque ratio}

\item{fun}{function, sert à calculer la statistique souhaitée
à partir de A et B}

\item{D}{Déviation max de la CKM}

\item{V}{Variance de la CKM}

\item{js}{Seuil interdiction de la CKM}

\item{betas}{Seuil des écarts entre R et R'}

\item{posterior}{Booléen: Si \code{TRUE}, c'est beta' qui est calculé.}

\item{parallel}{Booléen, si le calcul doit être parallélisé}

\item{max_cores, }{integer, nombre maximal de travaux à réaliser en parallèle}
}
\value{
un dataframe
}
\description{
Calcule P(|R-R'|>beta) pour chaque R=A/B d'un dataframe donné et pour
chaque beta fixé.
}
\details{
Si \code{posterior=FALSE}, le calcul repose sur l'approche a priori,
c'est-à-dire que le ratio fourni (A/B) est le ratio réel/original.
Sinon, le calcul repose sur l'approche a posteriori,
c'est-à-dire que le ratio fourni (A/B) est le ratio issu de la perturbation
par CKM.
}
\examples{
test <- data.frame(
  A = sample(1:50, 10, replace = TRUE),
  B = sample(50:1000, 10, replace = TRUE)
)

fun = \(a,b){a/b * 100}
D = 10
V = 10
js = 4

# Approche a priori
res <- calculer_proba_ecart_ratio_sur_df(test, fun, D, V)

# Approche a posteriori
res_ap <- calculer_proba_ecart_ratio_sur_df(test, fun, D, V, posterior = TRUE)
}
