% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CardKMeans.R
\name{CardKMeans}
\alias{CardKMeans}
\title{The current function calculates the Cardinality KMeans (i.e. CKM) solution, without selecting the number of masking variables and clusters (they are assumed known)
Note that the current function is an internal function that should not be used by the end-users}
\usage{
CardKMeans(dataset, n.cluster, n.noisevar, num_starts_kmeans = 10)
}
\arguments{
\item{dataset}{the orginal dataset on which CKM operates}

\item{n.cluster}{the total number of clusters}

\item{n.noisevar}{the total number of masking variables}

\item{num_starts_kmeans}{the number of starts for the conventional KM analysis (note that in CKM, the conventional KM operates in the lower dimensions). The default value is 10}
}
\value{
a list of three elements. The first is the vector that indicates the cluster assignment, the second is a vector that contains the
indices of all signaling variables, and the third is the total value of between-cluster sum of squares
}
\description{
The current function calculates the Cardinality KMeans (i.e. CKM) solution, without selecting the number of masking variables and clusters (they are assumed known)
Note that the current function is an internal function that should not be used by the end-users
}
\examples{
ncluster <- 3
nnoisevar <- 100
CardKMeans(dataset, ncluster, nnoisevar)
}
