% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKMSelNo.R
\name{CKMSelNo}
\alias{CKMSelNo}
\title{The current function calculates the Cardinality KMeans (i.e. CKM) solution, without selecting the number of masking variables and clusters (they are assumed known)}
\usage{
CKMSelNo(dataset, n.cluster, n.noisevar, num_starts_kmeans = 10)
}
\arguments{
\item{dataset}{the orginal dataset on which CKM operates}

\item{n.cluster}{the total number of clusters}

\item{n.noisevar}{the total number of masking variables}

\item{num_starts_kmeans}{the number of starts for the conventional KM analysis (note that in CKM, the conventional KM operates in the lower dimensions). The default value is 10}
}
\value{
@return The function will return a ckm object that is the list of five elements. The first denotes the selected number of masking variables; the second includes all indicies of signaling variables; the third is a vector illustrating cluster assignment; the forth is the pre-determined or selected "optimal" number of clusters; the fifth is the original dataset.
}
\description{
The current function calculates the Cardinality KMeans (i.e. CKM) solution, without selecting the number of masking variables and clusters (they are assumed known)
}
\examples{
ncluster <- 3
nnoisevar <- 100
ckm.sel.no <- CKMSelNo(dataset, ncluster, nnoisevar)
}
