% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_tableau.R
\name{compute_margins}
\alias{compute_margins}
\title{Compute margins for a contingency table}
\usage{
compute_margins(inner_cells, margins, resp_var = NULL, marge_label)
}
\arguments{
\item{inner_cells}{data.table. Inner cells of the table}

\item{margins}{Character vector. Margin variables}

\item{resp_var}{Character. Response variable to aggregate}

\item{marge_label}{Character. Label for margin cells}
}
\value{
data.table with computed margins
}
\description{
Compute margins for a contingency table
}
\keyword{internal}
