% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSCA.R
\name{ComputeSCA}
\alias{ComputeSCA}
\title{The function computes the solution of a special configuration of Sparce PCA (called SParse Variable PCA; SVPCA).
Instead of imposing zeros on individual elements of a loading matrix, all elements of a certain role (i.e. variables) are imposed to zero}
\usage{
ComputeSCA(
  data,
  r,
  n_sparvar,
  rational = 1,
  num_starts = 10,
  MAXITER = 1000,
  stop_value = 0.001
)
}
\arguments{
\item{data}{the orginal dataset on which SVPCA operates}

\item{r}{number of components}

\item{n_sparvar}{number of variables that have all zero loadings}

\item{rational}{whether using rational starts (1, the default value) or random starts (0). The number of rational starts is fixed at 1. Otherwise, the number of starts is given by num_starts.}

\item{num_starts}{the number of random starts used when \code{rational == 0}. The default value is 10}

\item{MAXITER}{the maximum number of iterations. The default value is 1000}

\item{stop_value}{the convergence criteria. The default value if 1e-3}
}
\value{
a list that contains three elements: the total loss value, the score matrices as well as the loading matrices
}
\description{
The function computes the solution of a special configuration of Sparce PCA (called SParse Variable PCA; SVPCA).
Instead of imposing zeros on individual elements of a loading matrix, all elements of a certain role (i.e. variables) are imposed to zero
}
\examples{
ncluster <- 3
nnoisevar <- 100
ComputeSCA(dataset, n.cluster, n.noisevar)
}
