% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabuler_et_appliquer_ckm_sur_liste.R
\name{convertir_desc_table_en_liste}
\alias{convertir_desc_table_en_liste}
\title{Convertit un tableau de description en listes de variables}
\usage{
convertir_desc_table_en_liste(desc_tab, prefix = "tab")
}
\arguments{
\item{desc_tab}{data.frame de 3 colonnes décrivant les tableaux à construire}

\item{prefix}{character Préfixe à ajouter aux noms de tables fournis dans desc_tab}
}
\value{
listes
}
\description{
Convertit un tableau de description en listes de variables
}
\examples{
# example code

desc_tableaux <- data.frame(
  TAB = c(rep(1,3), rep(2,4)),
  VAR = c("DIPLOME", "SEXE", "AGE", "DIPLOME", "TYPE", "REG", "DEP"),
  HRC = c(rep(NA, 5), rep("GEO",2))
)
convertir_desc_table_en_liste(desc_tableaux)
}
