% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrice_transition.R
\name{creer_matrice_transition}
\alias{creer_matrice_transition}
\title{Wrapper de la fonction `ptable::create_cnt_ptable`}
\usage{
creer_matrice_transition(D, V, js = 0, ...)
}
\arguments{
\item{D}{Déviation (integer strictement positif)}

\item{V}{Variance du bruit (numeric strictement positif)}

\item{js}{Seuil des valeurs interdites (integer). Si `js=0` (défaut), seule la
valeur 0 sera interdite.}

\item{...}{Paramètres additionnels de la fonction `ptable::create_cnt_ptable`}
}
\value{
objet `ptable` contenant la matrice de transition
}
\description{
Wrapper de la fonction `ptable::create_cnt_ptable`
}
\examples{
library(ptable)
mat_trans <- creer_matrice_transition(D = 5, V = 2)
plot(mat_trans, type="d") |> print()
}
