% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrice_transition.R
\name{creer_table_perturbation}
\alias{creer_table_perturbation}
\title{Prépare la table de perturbation à partir de la matrice de transition}
\usage{
creer_table_perturbation(matrice_transition)
}
\arguments{
\item{matrice_transition}{objet créé par `creer_matrice_transition()`}
}
\value{
data.table  - la table de perturbation
}
\description{
Prépare la table de perturbation à partir de la matrice de transition
}
\examples{
mat_trans <- creer_matrice_transition(D = 5, V = 2)
tab_pert <- creer_table_perturbation(mat_trans)
}
