% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{distance_euclid}
\alias{distance_euclid}
\title{Distance Euclidienne}
\usage{
distance_euclid(o, p)
}
\arguments{
\item{o}{vecteur des valeurs originales}

\item{p}{vecteur des valeurs perturbées}
}
\value{
numeric
}
\description{
Distance Euclidienne
}
\examples{
distance_euclid(1:100, 11:110)
}
