% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{distance_hellinger}
\alias{distance_hellinger}
\title{Hellinger distance}
\usage{
distance_hellinger(o, p)
}
\arguments{
\item{o}{numeric vector. Original values}

\item{p}{numeric vector. Perturbed values}
}
\value{
numeric. Hellinger distance value
}
\description{
Calculates the Hellinger distance between original and perturbed count vectors.
}
\examples{
distance_hellinger(1:100, 11:110)
}
