% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{distance_manhattan}
\alias{distance_manhattan}
\title{Manhattan distance}
\usage{
distance_manhattan(o, p)
}
\arguments{
\item{o}{numeric vector. Original values}

\item{p}{numeric vector. Perturbed values}
}
\value{
numeric. Manhattan distance value
}
\description{
Calculates the Manhattan distance between original and perturbed count vectors.
}
\examples{
distance_manhattan(1:100, 11:110)
}
