% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{ecarts_absolus_moyens}
\alias{ecarts_absolus_moyens}
\title{Ecarts absolus moyens}
\usage{
ecarts_absolus_moyens(o, p)
}
\arguments{
\item{o}{vecteur des valeurs originales}

\item{p}{vecteur des valeurs perturbées}
}
\value{
numeric
}
\description{
Ecarts absolus moyens
}
\examples{
ecarts_absolus_moyens(1:100, 11:110)
}
