% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{ecarts_absolus_moyens_relatifs}
\alias{ecarts_absolus_moyens_relatifs}
\title{Mean relative absolute deviations in percentage}
\usage{
ecarts_absolus_moyens_relatifs(o, p)
}
\arguments{
\item{o}{numeric vector. Original values}

\item{p}{numeric vector. Perturbed values}
}
\value{
numeric. Mean relative absolute deviation in percentage
}
\description{
Calculates the mean relative absolute deviations between original and
perturbed values, expressed as a percentage.
}
\examples{
ecarts_absolus_moyens_relatifs(1:100, 11:110)
}
