% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimer_intervalle_confiance_ratio.R
\name{estimer_proba_precision_statistique}
\alias{estimer_proba_precision_statistique}
\title{Estimate probability of ratio deviation}
\usage{
estimer_proba_precision_statistique(
  A,
  B,
  fun = function(a, b) {
     a/b * 100
 },
  D,
  V,
  js = 0,
  ptab,
  betas = c(0, 1, 5, 10, 20),
  posterior = FALSE
)
}
\arguments{
\item{A}{Numerator value}

\item{B}{Denominator value}

\item{fun}{Function. Statistic calculation (default: a/b*100)}

\item{D}{Integer. Maximum deviation}

\item{V}{Numeric. Noise variance}

\item{js}{Integer. Sensitivity threshold}

\item{ptab}{data.frame. Transition probabilities from ptable}

\item{betas}{Numeric vector. Precision thresholds to evaluate}

\item{posterior}{Logical. Use posterior approach? (default: FALSE)}
}
\value{
data.frame with beta probabilities
}
\description{
Calculates P(|R - R'| > \eqn{\beta}) for a statistic R = f(A, B) given CKM parameters.
}
\examples{
ptab <- ptable::create_cnt_ptable(D = 15, V = 30.1, js = 0)@pTable |>
  as.data.frame()

# Alpha value for the original statistic = 10/15 and different beta values
estimer_proba_precision_statistique(A=10,B=15,D = 15, V = 30.1, js = 0, ptab = ptab)

# Alpha value for the perturbed statistic = 10/15 and different beta' values
estimer_proba_precision_statistique(
  A=10,B=15,
  D = 15, V = 30.1, js = 0,
  ptab = ptab,
  posterior = TRUE
)

# For a ratio evolution
estimer_proba_precision_statistique(
  A=10,B=15,
  fun = \(a,b){(b/a - 1)*100},
  D = 15, V = 30.1, js = 0,
  ptab = ptab,
  posterior = TRUE
)
}
