% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_risque.R
\name{get_deviation_set}
\alias{get_deviation_set}
\title{Calculate deviation set for a given original value}
\usage{
get_deviation_set(i, D, js = 0)
}
\arguments{
\item{i}{integer. Original value for which to calculate the deviation set (must be non-negative)}

\item{D}{integer. Deviation parameter of the CKM (must be strictly positive)}

\item{js}{integer. Maximum forbidden value after perturbation (must be non-negative, default: 0)}
}
\value{
integer vector. Vector of possible perturbed values if the original value is i
}
\description{
Calculates the set of possible perturbed values for an original value i,
given the CKM parameters D and js.
}
\examples{
get_deviation_set(1, 5) # expected: 0:6
get_deviation_set(1, 5, 2) # expected: c(0,3:6)
get_deviation_set(0, 5, 2) # expected: 0
get_deviation_set(5, 5, 2) # expected: c(0,3:10)
}
