% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{globals}
\alias{globals}
\title{Global variables declaration}
\description{
This file declares global variables used throughout the package to avoid
R CMD check warnings about undefined global variables when using non-standard evaluation.
}
