% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{mean_absolute_deviation}
\alias{mean_absolute_deviation}
\title{Mean absolute deviations}
\usage{
mean_absolute_deviation(o, p)
}
\arguments{
\item{o}{numeric vector. Original values}

\item{p}{numeric vector. Perturbed values}
}
\value{
numeric. Mean absolute deviation
}
\description{
Calculates the mean absolute deviations between original and perturbed values.
}
\examples{
mean_absolute_deviation(1:100, 11:110)
}
