% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{mean_relative_absolute_deviation}
\alias{mean_relative_absolute_deviation}
\title{Mean relative absolute deviations in percentage}
\usage{
mean_relative_absolute_deviation(o, p)
}
\arguments{
\item{o}{numeric vector. Original values}

\item{p}{numeric vector. Perturbed values}
}
\value{
numeric. Mean relative absolute deviation in percentage
}
\description{
Calculates the mean relative absolute deviations between original and
perturbed values, expressed as a percentage.
}
\examples{
mean_relative_absolute_deviation(1:100, 11:110)
}
