% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_RU.R
\name{mesurer_RU}
\alias{mesurer_RU}
\title{Mesurer le risque et l'utilité d'un seul scénario}
\usage{
mesurer_RU(
  df,
  cat_vars,
  hrc_vars = NULL,
  D,
  V,
  js,
  confident,
  gv = 50,
  pv = 20,
  seed = NULL
)
}
\arguments{
\item{df}{data.frame or data.table}

\item{cat_vars}{vector of categorical variables but not hierarchical}

\item{hrc_vars}{named list (name = VAR final name, value = VAR current names)}

\item{D}{Déviation (integer strictement positif)}

\item{V}{Variance du bruit (numeric strictement positif)}

\item{js}{Seuil des valeurs interdites (integer). Si `js=0` (défaut), seule la
valeur 0 sera interdite.}

\item{confident}{\code{integer} Seuil (officiel) de confidentialité}

\item{gv}{\code{integer} seuil désignant les grands comptages}

\item{pv}{\code{integer} seuil désignant les petits comptages}

\item{seed}{\code{integer} numéro de graine aléatoire. Si \code{NULL},
la graine aléatoire utilisée est celle du programme parent.}
}
\value{
\code{data.frame} avec mesure de risque et mesure d'utilité pour
le paramétrage choisi.
}
\description{
`mesurer_RU()` mesure le risque et l'utilité d'un seul jeu de paramètres
à partir d'un tableau, sur un seul jeu de clés généré aléatoirement.
}
\section{La graine aléatoire}{


Le résultat sera identique si l'utilisateur choisi d'ajouter l'argument
`seed=123` ou bien opte pour fixer la graine par le classique `set.seed(123)`
dans son programme en amont de l'appel à la fonction.

Si les deux sont renseignés, c'est l'argument interne de la fonction qui prévaut.
}

\examples{
data("dtest")
set.seed(123)

res_RU <- mesurer_RU(
  df = dtest,
  cat_vars = c("REG", "DIPLOME", "SEXE", "AGE"),
  D = 10, V = 15, js = 4,
  confident = 10
)

}
