% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_RU.R
\name{mesurer_RUs}
\alias{mesurer_RUs}
\title{Comparer le risque et l'utilité de plusieurs scénarios}
\usage{
mesurer_RUs(
  df,
  cat_vars,
  hrc_vars = NULL,
  parametres,
  confident,
  gv = 50,
  pv = 20,
  seed = NULL
)
}
\arguments{
\item{df}{data.frame or data.table}

\item{cat_vars}{vector of categorical variables but not hierarchical}

\item{hrc_vars}{named list (name = VAR final name, value = VAR current names)}

\item{parametres}{\code{data.frame}}

\item{confident}{\code{integer} Seuil (officiel) de confidentialité}

\item{gv}{\code{integer} seuil désignant les grands comptages}

\item{pv}{\code{integer} seuil désignant les petits comptages}

\item{seed}{\code{integer} numéro de graine aléatoire. Si \code{NULL},
la graine aléatoire utilisée est celle du programme parent.}
}
\value{
\code{data.frame} de taille \code{nrow(parametres)}
}
\description{
`mesurer_RUs()` mesure le risque et l'utilité de
plusieurs jeux de paramètres à partir d'un tableau,
sur un seul jeu de clés généré aléatoirement.
}
\section{La graine aléatoire}{


Avec cette fonction, il est conseillé d'utiliser l'argument `seed=` si on souhaite
que tous les scénarios bénéficient de la même graine aléatoire et que les résultats
obtenus permettent de comparer les scénarios équiatblement.
}

\examples{
parametres <- construire_table_parametres(c(10,15), c(10,20), js = 5)
res_RUs <- mesurer_RUs(
  df = dtest,
  cat_vars = c("REG", "DIPLOME", "SEXE", "AGE"),
  parametres = parametres,
  confident = 10,
  seed = 1234
)
}
