% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCKM.R
\name{plot.ckm}
\alias{plot.ckm}
\title{The function plots the results of CKM: the cluster centroids of all signaling variables}
\usage{
\method{plot}{ckm}(ckm.object, var.ind = NULL)
}
\arguments{
\item{ckm.object}{an object of the S3 class "ckm", which could eitehr be the output of the function CKMSelVar (without the selection of number of clusters) or of the function CKMselAll (with the selection of nunber of clusters)}

\item{var.ind}{a vector of length J (i.e. the number of variables) that indicates the categories of variables}
}
\value{
a plot with x axis representing various clusters and y axis referring to the centroids of signaling variables
}
\description{
The function plots the results of CKM: the cluster centroids of all signaling variables
}
\examples{
ckm.sel.no <- CKMSelNo(dataset, ncluster, nnoisevar)
ckm.sel.var <- CKMSelVar(dataset, ncluster)
ckm.sel.all <- CKMSelAll(dataset, maxcluster)
plot(ckm.sel.no)
plot(ckm.sel.var)
plot(ckm.sel.all)


}
