% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrice_transition.R
\name{prepare_perturbation_table}
\alias{prepare_perturbation_table}
\title{Create perturbation table from transition matrix}
\usage{
prepare_perturbation_table(matrice_transition)
}
\arguments{
\item{matrice_transition}{ptable object. Object created by create_transition_matrix()}
}
\value{
data.table containing the perturbation table with columns i, v, p_int_lb, p_int_ub
}
\description{
Prepares a perturbation lookup table from a transition matrix object
for efficient application of the Cell Key Method.
}
\examples{
\dontrun{
mat_trans <- create_transition_matrix(D = 5, V = 2)
tab_pert <- prepare_perturbation_table(mat_trans)
}
}
