% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_tableau.R
\name{tabulate_cnt_micro_data}
\alias{tabulate_cnt_micro_data}
\title{Construit la table des effectifs en croisant l'ensemble des variables catégorielles
et en construisant l'ensemble des marges possibles. La fonction construit
également la clé aléatoire de la cellule à partir des clés individuelles.}
\usage{
tabulate_cnt_micro_data(
  df,
  rk_var = "rkey",
  cat_vars = NULL,
  hrc_vars = NULL,
  marge_label = "Total"
)
}
\arguments{
\item{df}{data.frame or data.table}

\item{rk_var}{(character) nom de la variable des clés individuelles}

\item{cat_vars}{vector of categorical variables but not hierarchical}

\item{hrc_vars}{named list (name = VAR final name, value = VAR current names)}

\item{marge_label}{label of margins (applied to all cat and hrc variables)}
}
\value{
a tibble. La variable de comptage est nommée `nb_obs`, la clé de la cellule
est dénommée `ckey`.
}
\description{
Construit la table des effectifs en croisant l'ensemble des variables catégorielles
et en construisant l'ensemble des marges possibles. La fonction construit
également la clé aléatoire de la cellule à partir des clés individuelles.
}
\examples{
library(data.table)

data("dtest")

tab_comptage <- tabulate_cnt_micro_data(
  df = dtest,
  rk_var = NULL,
  cat_vars = c("DIPLOME", "SEXE", "AGE"),
  hrc_vars = list(GEO = c("REG", "DEP")),
  marge_label = "Total"
)
str(tab_comptage)
}
