% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabuler_et_appliquer_ckm.R
\name{tabuler_et_appliquer_ckm}
\alias{tabuler_et_appliquer_ckm}
\title{Cette fonction construit le tableau et applique la CKM.}
\usage{
tabuler_et_appliquer_ckm(
  df,
  rk_var = "rkey",
  cat_vars,
  hrc_vars = NULL,
  marge_label = "Total",
  D,
  V,
  js = 0,
  i_risk = 1:4,
  j_risk = 1:4,
  ...
)
}
\arguments{
\item{df}{data.frame or data.table}

\item{rk_var}{(character) nom de la variable des clés individuelles}

\item{cat_vars}{vector of categorical variables but not hierarchical}

\item{hrc_vars}{named list (name = VAR final name, value = VAR current names)}

\item{marge_label}{label of margins (applied to all cat and hrc variables)}

\item{D}{Déviation (integer strictement positif)}

\item{V}{Variance du bruit (numeric strictement positif)}

\item{js}{Seuil des valeurs interdites (integer). Si `js=0` (défaut), seule la
valeur 0 sera interdite.}

\item{...}{Paramètres additionnels de la fonction `ptable::create_cnt_ptable`}
}
\value{
liste comprenant la table perturbée et la matrice de transition
}
\description{
Cette fonction construit le tableau et applique la CKM.
}
\examples{
data("dtest")
dtest_avec_cles <- construire_cles_indiv(dtest, 40889)

res_ckm <- tabuler_et_appliquer_ckm(
  df = dtest_avec_cles,
  cat_vars = c("DIPLOME", "SEXE", "AGE"),
  hrc_vars = list(GEO = c("REG", "DEP")),
  marge_label = "Total",
  D = 10, V = 15, js = 4,
  j_risk = 0
)
}
