% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester_matrices.R
\name{tester_matrices}
\alias{tester_matrices}
\title{Teste la construction des matrices de
transition pour différentes valeurs de variance à D et js fixés}
\usage{
tester_matrices(D, js = 0, Vmin = 0, Vmax = 30, precision = 1)
}
\arguments{
\item{D}{\code{integer} déviation}

\item{js}{\code{integer} seuil des valeurs sensibles}

\item{Vmin}{\code{numeric} Variance minimum à tester}

\item{Vmax}{\code{numeric} Variance maximum à tester}

\item{precision}{\code{numeric} Précision}
}
\value{
\code{NULL} si aucune solution n'existe dans l'intervalle demandé
\code{numeric} Valeur de la variance qui rend la matrice constructible
}
\description{
Teste la construction des matrices de
transition pour différentes valeurs de variance à D et js fixés
}
\examples{
tester_matrices(5, 1)
tester_matrices(10, js = 9)
tester_matrices(10, js = 9, Vmin=25, Vmax=30, precision=0.5)
}
