% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiser_distributions.R
\name{visualiser_distribution}
\alias{visualiser_distribution}
\title{Visualisation de distributions de probabilités
issues de plusieurs scénarios}
\usage{
visualiser_distribution(D = NULL, V = NULL, DV = NULL, precision = 5)
}
\arguments{
\item{D}{Déviation (integer strictement positif)}

\item{V}{Variance du bruit (numeric strictement positif)}

\item{precision}{\code{integer}}
}
\value{
cowplot object
}
\description{
Visualisation de distributions de probabilités
issues de plusieurs scénarios
}
\examples{
visualiser_distribution(
  D = c(11, 15),
  V = c(10, 30)
)
visualiser_distribution(
  DV = data.frame(D = c(11, 15), V = c(10, 30))
)
}
