% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appliquer_ckm.R
\name{appliquer_ckm}
\alias{appliquer_ckm}
\title{Appliquer une CKM sur un tableau}
\usage{
appliquer_ckm(
  tab_data,
  cnt_var = "nb_obs",
  ck_var = "ckey",
  D,
  V,
  js = 0,
  I = NULL,
  J = NULL,
  ...
)
}
\arguments{
\item{tab_data}{objet retourné par `tabulate_cnt_micro_data`  (data.frame ou liste)}

\item{cnt_var}{(character) nom de la variable de comptage (fréquences, effectifs)}

\item{ck_var}{(character) nom de la variable contenant la clé aléatoire de la cellule
Il s'agit nécessairement d'un nombre décimal compris entre 0 et 1. La clé
d'une cellule doit correspondre à la partie décimale de la somme des
clés des individus qui composent la cellule.}

\item{D}{Déviation (integer strictement positif)}

\item{V}{Variance du bruit (numeric strictement positif)}

\item{js}{Seuil des valeurs interdites (integer). Si `js=0` (défaut), seule la
valeur 0 sera interdite.}

\item{I}{vecteur d'entiers (valeurs originales)}

\item{J}{vecteur d'entiers (valeurs perturbées)}

\item{...}{Paramètres additionnels de la fonction `ptable::create_cnt_ptable`}
}
\value{
liste comprenant la table perturbée, la matrice de transition, la mesure du
risque (NULL si les fréquences empiriques ne sont pas fournies) et la mesure
de l'utilité.
}
\description{
Appliquer une CKM sur un tableau
}
\examples{
library(dplyr)
data("dtest")
set.seed(8245)
dtest_avec_cles <- construire_cles_indiv(dtest)

tab_avant <- tabulate_cnt_micro_data(
  df = dtest_avec_cles,
  cat_vars = c("DIPLOME", "SEXE", "AGE"),
  hrc_vars = list(GEO = c("REG", "DEP")),
  marge_label = "Total",
  freq_empiriq = TRUE #pour pouvoir mesurer le risque
)

res_ckm <- appliquer_ckm(tab_avant, D = 5, V = 2)
str(res_ckm, max.level = 2)

tab_avant2 <- tabulate_cnt_micro_data(
  df = dtest_avec_cles |> mutate(NUM = 12),
  cat_vars = c("DIPLOME", "SEXE", "AGE"),
  hrc_vars = list(GEO = c("REG", "DEP")),
  num_var = "NUM",
  marge_label = "Total",
  freq_empiriq = TRUE #pour pouvoir mesurer le risque
)

res_ckm2 <- appliquer_ckm(tab_avant2, cnt_var = "num_tot", D = 5, V = 2)
head(res_ckm2$tab)
}
