% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appliquer_ckm.R
\name{appliquer_ckm}
\alias{appliquer_ckm}
\title{Appliquer une CKM sur un tableau}
\usage{
appliquer_ckm(tab_data, cnt_var = "nb_obs", ck_var = "ckey", D, V, js = 0, ...)
}
\arguments{
\item{tab_data}{tableau agrégé}

\item{cnt_var}{(character) nom de la variable de comptage (fréquences, effectifs)}

\item{ck_var}{(character) nom de la variable contenant la clé aléatoire de la cellule
Il s'agit nécessairement d'un nombre décimal compris entre 0 et 1. La clé
d'une cellule doit correspondre à la partie décimale de la somme des
clés des individus qui composent la cellule.}

\item{D}{Déviation (integer strictement positif)}

\item{V}{Variance du bruit (numeric strictement positif)}

\item{js}{Seuil des valeurs interdites (integer). Si `js=0` (défaut), seule la
valeur 0 sera interdite.}

\item{...}{Paramètres additionnels de la fonction `ptable::create_cnt_ptable`}
}
\value{
liste comprenant la table perturbée et la matrice de transition
}
\description{
Appliquer une CKM sur un tableau
}
\examples{
data("dtest")
dtest_avec_cles <- construire_cles_indiv(dtest, 40889)

tab_avant <- tabulate_cnt_micro_data(
  df = dtest_avec_cles,
  cat_vars = c("DIPLOME", "SEXE", "AGE"),
  hrc_vars = list(GEO = c("REG", "DEP")),
  marge_label = "Total"
)

res_ckm <- appliquer_ckm(tab_avant, D = 5, V = 2)
str(res_ckm$tab)
}
