% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_risque.R
\name{calculer_ensemble_possibles}
\alias{calculer_ensemble_possibles}
\title{Calcule l'ensemble des possibles de j étant donnés les paramètres D et js
de la matrice de transition}
\usage{
calculer_ensemble_possibles(j, D, js = 0)
}
\arguments{
\item{j}{valeur perturbée dont on souhaite calculer l'intervalle des possibles}

\item{D}{la déviation de la CKM}

\item{js}{Maximum des valeurs interdites après perturbation}
}
\value{
vecteur des valeurs originales possibles si la valeur perturbée est j,
calculé en fonction de D et js. `NULL` si j > 0 et j <= js
}
\description{
Calcule l'ensemble des possibles de j étant donnés les paramètres D et js
de la matrice de transition
}
\examples{

calculer_ensemble_possibles(1, 5) #attendu 1:6
calculer_ensemble_possibles(1, 5, 2) #attendu NULL
calculer_ensemble_possibles(0, 5, 2) #attendu 0:5
calculer_ensemble_possibles(5, 5, 2) #attendu 1:10
}
