% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_risque.R
\name{calculer_frequences_empiriques}
\alias{calculer_frequences_empiriques}
\title{Calcule les fréquences empiriques des comptages à partir d'un tableau agrégé
construit avec les fonctions `tabulate_cnt_micro_data` ou `appliquer_ckm` ou
`tabuler_et_appliquer_ckm`.}
\usage{
calculer_frequences_empiriques(tableau, cat_vars, hrc_vars)
}
\arguments{
\item{tableau}{tableau généré avec les fonctions `tabulate_cnt_micro_data` ou
`appliquer_ckm` ou `tabuler_et_appliquer_ckm`}

\item{cat_vars}{vector of categorical variables but not hierarchical}

\item{hrc_vars}{named list (name = VAR final name, value = VAR current names)}
}
\value{
data.frame avec 3 colonnes:
- i = comptage
- N = nb d'apparitions du comptage
- p_hat = fréquence empirique du comptage
}
\description{
Calcule les fréquences empiriques des comptages à partir d'un tableau agrégé
construit avec les fonctions `tabulate_cnt_micro_data` ou `appliquer_ckm` ou
`tabuler_et_appliquer_ckm`.
}
\examples{
library(ptable)
library(dplyr)

data("dtest")

cat_vars1 = c("DEP", "DIPLOME", "SEXE", "AGE")
hrc_vars1 = NULL
tab_comptage1 <- tabulate_cnt_micro_data(
  df = dtest, rk = NULL,
  cat_vars = cat_vars1,
  marge_label = "Total"
)
p_hat1 <- calculer_frequences_empiriques(tab_comptage1, cat_vars1, hrc_vars1)

cat_vars2 = c("DIPLOME", "SEXE", "AGE")
hrc_vars2 = list(GEO = c("REG","DEP"), TYPES = c("TYPE","TYPE2"))
tab_comptage2 <- tabulate_cnt_micro_data(
  df = dtest, rk = NULL,
  cat_vars = cat_vars2,
  hrc_vars = hrc_vars2,
  marge_label = "Total"
)
p_hat2 <- calculer_frequences_empiriques(tab_comptage2, cat_vars2, hrc_vars2)
}
\keyword{internal}
