% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_risque.R
\name{calculer_frequences_empiriques}
\alias{calculer_frequences_empiriques}
\title{Calcule les fréquences empiriques des comptages à partir d'un tableau agrégé
construit avec les fonctions `tabulate_cnt_micro_data` ou `appliquer_ckm` ou
`tabuler_et_appliquer_ckm`.}
\usage{
calculer_frequences_empiriques(tableau, hierarchies = NULL)
}
\arguments{
\item{tableau}{tableau généré avec les fonctions `tabulate_cnt_micro_data` ou
`appliquer_ckm` ou `tabuler_et_appliquer_ckm`}

\item{hierarchies}{liste de vecteurs - description des emboîtements éventuelles
des variables catégorielles de tableau entre elles}
}
\value{
data.frame avec 3 colonnes:
- i = comptage
- N = nb d'apparitions du comptage
- p_hat = fréquence empirique du comptage
}
\description{
Calcule les fréquences empiriques des comptages à partir d'un tableau agrégé
construit avec les fonctions `tabulate_cnt_micro_data` ou `appliquer_ckm` ou
`tabuler_et_appliquer_ckm`.
}
\examples{
library(ptable)
library(dplyr)

data("dtest")

tab_comptage1 <- tabulate_cnt_micro_data(
  df = dtest, rk = NULL,
  cat_vars = c("DEP", "DIPLOME", "SEXE", "AGE"),
  marge_label = "Total"
)
p_hat1 <- calculer_frequences_empiriques(tab_comptage1)

tab_comptage2 <- tabulate_cnt_micro_data(
  df = dtest, rk = NULL,
  cat_vars = c("DEP", "DIPLOME", "SEXE", "AGE"),
  marge_label = "Total"
)
p_hat <- calculer_frequences_empiriques(tab_comptage1)
}
