% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabuler_et_appliquer_ckm.R
\name{check_inputs_tabulate}
\alias{check_inputs_tabulate}
\title{Check Inputs for Tabulation Function}
\usage{
check_inputs_tabulate(
  df,
  rk_var,
  cat_vars,
  hrc_vars,
  num_var,
  marge_label,
  D,
  V,
  js,
  I,
  J,
  stack
)
}
\arguments{
\item{df}{A data frame containing the data to be tabulated.}

\item{rk_var}{A character string specifying the name of the individual key variable. Must be a single character string and present in \code{df}.}

\item{cat_vars}{A character vector of categorical variable names. All must be present in \code{df}. Can be \code{NULL}.}

\item{hrc_vars}{A list of hierarchical variable names. All must be present in \code{df}. Can be \code{NULL}.}

\item{num_var}{A character string specifying the name of the numerical variable. Must be present in \code{df}. Can be \code{NULL}.}

\item{marge_label}{A character string specifying the label for the margin. Must be a single character string.}

\item{D}{integer. Deviation parameter (must be strictly positive)}

\item{V}{numeric. Noise variance (must be strictly positive)}

\item{js}{integer. Threshold for sensitive values (default: 0). If js=0, only value 0 will be forbidden}

\item{I}{integer vector. Original values to consider}

\item{J}{integer vector. Perturbed values to consider}

\item{stack}{Named list of parameters (D, V) to produce a stacked matrix
with these parameters applied only on the large counts. js is set to 0 for this part.}
}
\value{
None. The function is called for its side effects (throws an error if any assertion fails).
}
\description{
Validates the inputs provided to the tabulation function, ensuring that all required arguments are present and correctly specified.
}
\details{
This function uses assertions to check the validity of each input argument. If any check fails, an informative error message is provided.
}
\examples{
\dontrun{
check_inputs_tabulate(
  df = my_data,
  rk_var = "id",
  cat_vars = c("gender", "region"),
  hrc_vars = list("region"),
  num_var = "income",
  marge_label = "Total",
  D = 1,
  V = 2,
  js = 0,
  I = NULL,
  J = NULL
)
}

}
\keyword{internal}
