% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckm-package.R
\docType{package}
\name{ckm-package}
\alias{ckm}
\alias{ckm-package}
\title{ckm: Applying the Cell Key Method to Protect Tabular Data}
\description{
The Cell Key Method is a method for protecting sensitive data in tabular form by applying random keys to cells. The package implements the method and provides tools to choose the best parameters to ensure the balance between data utility and privacy protection. See Enderle et al. (2020) \doi{10.1007/978-3-030-57521-2_11}, Jamme (2025) \doi{10.5281/zenodo.21930561} and Jamme (2026) \doi{10.13140/RG.2.2.19275.55840} for further details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/InseeFrLab/ckm}
  \item \url{https://inseefrlab.github.io/ckm/}
  \item Report bugs at \url{https://github.com/InseeFrLab/ckm/issues}
}

}
\author{
\strong{Maintainer}: Julien Jamme \email{julien.jamme@insee.fr}

Authors:
\itemize{
  \item Julien Jamme \email{julien.jamme@insee.fr}
}

Other contributors:
\itemize{
  \item  Institut National de la Statistique et des Études Économiques [copyright holder]
}

}
\keyword{internal}
