% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckm-package.R
\docType{package}
\name{ckm-package}
\alias{ckm}
\alias{ckm-package}
\title{ckm: Appliquer la méthode des clés aléatoires (Cell Key Method) pour protéger des données tabulées}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Julien Jamme \email{julien.jamme@insee.fr}

Other contributors:
\itemize{
  \item  Institut National de la Statistique et des Études Économiques [copyright holder]
}

}
\keyword{internal}
