% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_tableau.R
\name{compute_margins}
\alias{compute_margins}
\title{Fonction construisant les marges pour un croisement donné}
\usage{
compute_margins(inner_cells, margins, resp_var = NULL, marge_label)
}
\arguments{
\item{inner_cells}{data.table}

\item{margins}{character}

\item{resp_var}{character}

\item{marge_label}{character}
}
\value{
data.table
}
\description{
Fonction construisant les marges pour un croisement donné
}
