% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construire_cles_indiv.R
\name{construire_cles_indiv}
\alias{construire_cles_indiv}
\title{Generate individual record keys}
\usage{
construire_cles_indiv(microdata, nb_decim = NULL)
}
\arguments{
\item{microdata}{data.frame. Input microdata}

\item{nb_decim}{Integer. Key precision (auto-calculated if NULL)}
}
\value{
data.table with added 'rkey' column
}
\description{
Adds random decimal keys to microdata for Cell Key Method implementation.
}
\examples{
data("dtest")
set.seed(123)
dtest_avec_cles <- construire_cles_indiv(dtest)

}
