% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construire_table_parametres.R
\name{construire_table_parametres}
\alias{construire_table_parametres}
\title{Construit un data.frame de tous les jeux de paramètres possibles}
\usage{
construire_table_parametres(Ds, Vs, jss = 0)
}
\arguments{
\item{Ds}{\code{integer} vecteur de déviations}

\item{Vs}{\code{numeric} vecteur de variance}

\item{jss}{\code{integer} vecteur de seuil de sensibilité}
}
\value{
\code{data.frame}
}
\description{
Construit un data.frame de tous les jeux de paramètres possibles
}
\examples{
construire_table_parametres(c(10,15), c(10,20), js = 5)
}
