% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{distance_euclid}
\alias{distance_euclid}
\title{Euclidean distance}
\usage{
distance_euclid(o, p)
}
\arguments{
\item{o}{numeric vector. Original values}

\item{p}{numeric vector. Perturbed values}
}
\value{
numeric. Euclidean distance value
}
\description{
Calculates the Euclidean distance between original and perturbed count vectors.
}
\examples{
distance_euclid(1:100, 11:110)
}
