% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{distance_hellinger}
\alias{distance_hellinger}
\title{Distance de Hellinger}
\usage{
distance_hellinger(o, p)
}
\arguments{
\item{o}{vecteur des valeurs originales}

\item{p}{vecteur des valeurs perturbées}
}
\value{
numeric
}
\description{
Distance de Hellinger
}
\examples{
distance_hellinger(1:100, 11:110)
}
