% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{distance_manhattan}
\alias{distance_manhattan}
\title{Distance de Manhattan}
\usage{
distance_manhattan(o, p)
}
\arguments{
\item{o}{vecteur des valeurs originales}

\item{p}{vecteur des valeurs perturbées}
}
\value{
numeric
}
\description{
Distance de Manhattan
}
\examples{
distance_manhattan(1:100, 11:110)
}
