% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{ecarts_absolus_moyens}
\alias{ecarts_absolus_moyens}
\title{Mean absolute deviations}
\usage{
ecarts_absolus_moyens(o, p)
}
\arguments{
\item{o}{numeric vector. Original values}

\item{p}{numeric vector. Perturbed values}
}
\value{
numeric. Mean absolute deviation
}
\description{
Calculates the mean absolute deviations between original and perturbed values.
}
\examples{
ecarts_absolus_moyens(1:100, 11:110)
}
