% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_utilite.R
\name{ecarts_absolus_moyens_relatifs}
\alias{ecarts_absolus_moyens_relatifs}
\title{Ecarts absolus relatifs moyens en %}
\usage{
ecarts_absolus_moyens_relatifs(o, p)
}
\arguments{
\item{o}{vecteur des valeurs originales}

\item{p}{vecteur des valeurs perturbées}
}
\value{
numeric (en %)
}
\description{
Ecarts absolus relatifs moyens en %
}
\examples{
ecarts_absolus_moyens_relatifs(1:100, 11:110)
}
