% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_risque.R
\name{mesurer_risque}
\alias{mesurer_risque}
\title{Mesure du risque en estimant les probabilités de transition inverses
Calcul les probabilités P(X=i|X'=j) où X désigne l'original et X' le perturbé}
\usage{
mesurer_risque(matrice_transition, freq, I, J)
}
\arguments{
\item{matrice_transition}{objet retourné par `creer_matrice_transition`}

\item{freq}{objet retourné par `calculer_frequences_empiriques`}

\item{I}{vecteur d'entiers (valeurs originales)}

\item{J}{vecteur d'entiers (valeurs perturbées)}
}
\value{
`data.frame` de 5 colonnes:
- `i`: valeur(s) prise(s) par X
- `j`: valeur(s) prise(s) par X'
- `pi_hat`: Estimation de `P( X = i )`
- `pij`: Probabilité de transition `P(X' = j | X = i )`
- `qij`: Probabilité de transition inverse `P( X = i | X' = j )`
}
\description{
Mesure du risque en estimant les probabilités de transition inverses
Calcul les probabilités P(X=i|X'=j) où X désigne l'original et X' le perturbé
}
\details{
Pour le code, voir les formules et algorithmes ici (ajouter lien vers document pdf).
}
\examples{
library(ptable)
library(dplyr)
mat_trans <- creer_matrice_transition(D = 5, V = 2)
data("dtest")

tab_comptage <- tabulate_cnt_micro_data(
  df = dtest, rk = NULL,
  cat_vars = c("DEP", "DIPLOME", "SEXE", "AGE"),
  marge_label = "Total",
  freq_empiriq = TRUE
)

# Ci-dessous calcul des probabilités de transition inverses P(X=i|X'=1) avec
#i qui prend toutes les valeurs entre 1 et 4 (et aussi l'ensemble).
mesurer_risque(mat_trans, tab_comptage$freq, 1:4, 1)

# Ci-dessous calcul des probabilités de transition inverses P(X=i|X'=j) avec
# i qui prend toutes les valeurs entre 1 et 4 (et aussi l'ensemble)
# et j qui prend toutes les valeurs entre 1 et 4 (et aussi l'ensemble).
mesurer_risque(mat_trans, tab_comptage$freq, 1:4, 1:4)
}
