% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesurer_RU.R
\name{mesurer_RUs}
\alias{mesurer_RUs}
\title{Compare risk and utility across multiple scenarios}
\usage{
mesurer_RUs(
  df,
  cat_vars,
  hrc_vars = NULL,
  parametres,
  confident,
  gv = 50,
  pv = 20,
  seed = NULL
)
}
\arguments{
\item{df}{data.frame. Input microdata}

\item{cat_vars}{Character vector. Categorical variables}

\item{hrc_vars}{Named list. Hierarchical variables}

\item{parametres}{data.frame. Parameter combinations to test with columns D, V, js}

\item{confident}{integer. Official confidentiality threshold}

\item{gv}{integer. Threshold defining large counts (default: 50)}

\item{pv}{integer. Threshold defining small counts (default: 20)}

\item{seed}{integer. Random seed number. If NULL, uses parent program's seed}
}
\value{
data.frame with nrow(parametres) rows containing risk and utility measures
}
\description{
Measures risk and utility for multiple sets of CKM parameters 
from a table, using one randomly generated set of record keys.
}
\section{Random seed}{

With this function, it is recommended to use the seed= argument if you want
all scenarios to benefit from the same random seed and ensure results
allow fair comparison between scenarios.
}

\examples{
\dontrun{
parametres <- construire_table_parametres(c(10,15), c(10,20), js = 5)
res_RUs <- mesurer_RUs(
  df = dtest,
  cat_vars = c("REG", "DIPLOME", "SEXE", "AGE"),
  parametres = parametres,
  confident = 10,
  seed = 1234
)
}

}
