% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabuler_et_appliquer_ckm_sur_liste.R
\name{tabuler_et_appliquer_ckm_liste}
\alias{tabuler_et_appliquer_ckm_liste}
\title{Cette fonction construit le tableau et applique la CKM.}
\usage{
tabuler_et_appliquer_ckm_liste(
  df,
  rk_var = "rkey",
  desc_tab,
  marge_label = "Total",
  prefix = "tab",
  D,
  V,
  js = 0,
  I = NULL,
  J = NULL,
  ...
)
}
\arguments{
\item{df}{data.frame or data.table}

\item{rk_var}{(character) nom de la variable des clés individuelles}

\item{desc_tab}{data.frame de 3 colonnes décrivant les tableaux à construire}

\item{marge_label}{label of margins (applied to all cat and hrc variables)}

\item{prefix}{character Préfixe à ajouter aux noms de tables fournis dans desc_tab}

\item{D}{Déviation (integer strictement positif)}

\item{V}{Variance du bruit (numeric strictement positif)}

\item{js}{Seuil des valeurs interdites (integer). Si `js=0` (défaut), seule la
valeur 0 sera interdite.}

\item{I}{vecteur d'entiers (valeurs originales)}

\item{J}{vecteur d'entiers (valeurs perturbées)}

\item{...}{Paramètres additionnels de la fonction `ptable::create_cnt_ptable`}
}
\value{
liste comprenant les éléments
\itemize{
\item \code{tab}: liste des tables (tibbles) construites sur chacune desquelles a été appliquée la CKM
\item \code{ptab}: La matrice de transition qui a servi au calcul
\item \code{risque}: Tibble regroupant l'ensemble des mesures de risque correspondant à chacune des tables
\item \code{utilite}: Tibble regroupant l'ensemble des mesures d'utilité correspondant à chacune des tables
}
}
\description{
Cette fonction construit le tableau et applique la CKM.
}
\details{
Le data.frame à fournir dans l'argument \code{desc_tab} doit être constitué de telle façon que:
\enumerate{
\item Le data.frame doit contenir les 3 colonnes suivantes:
}
\itemize{
\item \code{TAB}: Nom ou numéro de la table
\item \code{VAR}: Nom d'une variable de la table
\item \code{HRC}: Si la variable a une relation hiérarchique avec une autre des variables de la table
mentionner un nom pour cette hiérarchie. Sinon, \code{NA}
}
\enumerate{
\item Une ligne correspond à une variable d'une table donnée
\item Deux variables (ex: REGION et DEPARTEMENT) d'une même table ayant une relation hiérarchique
doivent recevoir la même valeur dans la colonne \code{HRC}. Ces variables
doivent être renseignées dans l'ordre décroissant de la hiérarchie,
cad du niveau le plus large (ex: REGION) au niveau le plus fin (ex: DEPARTEMENT).
}

Voir la partie exemple pour une illustration.
}
\examples{
data("dtest")
set.seed(123)
dtest_avec_cles <- construire_cles_indiv(dtest)

# On se base sur deux tableaux:
# tab1: DIPLOME * SEXE * AGE
# tab1: DIPLOME * TYPE * REG * DEP, où REG > DEP

desc_tableaux <- data.frame(
  TAB = c(rep(1,3), rep(2,4)),
  VAR = c("DIPLOME", "SEXE", "AGE", "DIPLOME", "TYPE", "REG", "DEP"),
  HRC = c(rep(NA, 5), rep("GEO",2))
)
desc_tableaux

res_ckm <- tabuler_et_appliquer_ckm_liste(
  df = dtest_avec_cles,
  desc_tab = desc_tableaux,
  marge_label = "Total",
  D = 10, V = 15, js = 4
)
}
