% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester_matrices.R
\name{tester_matrices}
\alias{tester_matrices}
\title{Teste la construction des matrices de
transition pour différentes valeurs de variance à D et js fixés}
\usage{
tester_matrices(D, js = 0, Vmin = 0, Vmax = 30, Vseq = 5)
}
\arguments{
\item{D}{integer déviation}

\item{js}{integer seuil des valeurs sensibles}

\item{Vmin}{numeric Variaince minimum à tester}

\item{Vmax}{numeric Variance maximum à tester}

\item{Vseq}{integer pas d'avancement}
}
\value{
data.frame précisant si chaque matrice a pu être construite ou non
}
\description{
Teste la construction des matrices de
transition pour différentes valeurs de variance à D et js fixés
}
\examples{
tester_matrices(5, 1)
tester_matrices(10, js = 9)
tester_matrices(10, js = 9, Vmin=25, Vmax=30, Vseq=1)
}
